+++ cpp/src/io/parquet/experimental/hybrid_scan_helpers.hpp
@@ -139,16 +139,6 @@ class aggregate_reader_metadata : public aggregate_reader_metadata_base {
                            type_id timestamp_type_id);
 
   /**
-   * @brief Filters and reduces down to a selection of row groups
-   *
-   * @param row_group_indices Lists of row groups to read, one per source
-   *
-   * @return A tuple of total number of rows to read and list of row group info
-   */
-  [[nodiscard]] std::tuple<size_type, std::vector<row_group_info>> select_row_groups(
-    host_span<std::vector<size_type> const> row_group_indices);
-
-  /**
    * @brief Filter the row groups with statistics based on predicate filter
    *
    * @param row_group_indices Input row groups indices
