+++ cpp/tests/io/experimental/hybrid_scan_predicate_pushdown_test.cpp
@@ -14,243 +14,22 @@
  * limitations under the License.
  */
 
-#include "parquet_common.hpp"
+#include "hybrid_scan_common.hpp"
 
 #include <cudf_test/base_fixture.hpp>
-#include <cudf_test/column_wrapper.hpp>
-#include <cudf_test/io_metadata_utilities.hpp>
-#include <cudf_test/iterator_utilities.hpp>
-#include <cudf_test/table_utilities.hpp>
 
-#include <cudf/column/column.hpp>
-#include <cudf/concatenate.hpp>
 #include <cudf/io/experimental/hybrid_scan.hpp>
 #include <cudf/io/parquet.hpp>
 #include <cudf/io/text/byte_range_info.hpp>
 #include <cudf/stream_compaction.hpp>
-#include <cudf/table/table.hpp>
 #include <cudf/table/table_view.hpp>
-#include <cudf/transform.hpp>
 #include <cudf/utilities/default_stream.hpp>
 #include <cudf/utilities/memory_resource.hpp>
 #include <cudf/utilities/span.hpp>
-#include <cudf/utilities/traits.hpp>
 
 #include <src/io/parquet/parquet_gpu.hpp>
 
-// Base test fixture for tests
-struct ParquetExperimentalReaderTest : public cudf::test::BaseFixture {};
-
 namespace {
-
-/**
- * @brief Fetches a host span of Parquet footer bytes from the input buffer span
- *
- * @param buffer Input buffer span
- * @return A host span of the footer bytes
- */
-cudf::host_span<uint8_t const> fetch_footer_bytes(cudf::host_span<uint8_t const> buffer)
-{
-  using namespace cudf::io::parquet;
-
-  constexpr auto header_len = sizeof(file_header_s);
-  constexpr auto ender_len  = sizeof(file_ender_s);
-  size_t const len          = buffer.size();
-
-  auto const header_buffer = cudf::host_span<uint8_t const>(buffer.data(), header_len);
-  auto const header        = reinterpret_cast<file_header_s const*>(header_buffer.data());
-  auto const ender_buffer =
-    cudf::host_span<uint8_t const>(buffer.data() + len - ender_len, ender_len);
-  auto const ender = reinterpret_cast<file_ender_s const*>(ender_buffer.data());
-  CUDF_EXPECTS(len > header_len + ender_len, "Incorrect data source");
-  constexpr uint32_t parquet_magic = (('P' << 0) | ('A' << 8) | ('R' << 16) | ('1' << 24));
-  CUDF_EXPECTS(header->magic == parquet_magic && ender->magic == parquet_magic,
-               "Corrupted header or footer");
-  CUDF_EXPECTS(ender->footer_len != 0 && ender->footer_len <= (len - header_len - ender_len),
-               "Incorrect footer length");
-
-  return cudf::host_span<uint8_t const>(buffer.data() + len - ender->footer_len - ender_len,
-                                        ender->footer_len);
-}
-
-/**
- * @brief Fetches a host span of Parquet page index bytes from the input buffer span
- *
- * @param buffer Input buffer span
- * @param page_index_bytes Byte range of page index to fetch
- * @return A host span of the page index bytes
- */
-cudf::host_span<uint8_t const> fetch_page_index_bytes(
-  cudf::host_span<uint8_t const> buffer, cudf::io::text::byte_range_info const page_index_bytes)
-{
-  return cudf::host_span<uint8_t const>(
-    reinterpret_cast<uint8_t const*>(buffer.data()) + page_index_bytes.offset(),
-    page_index_bytes.size());
-}
-
-/**
- * @brief Fetches a list of byte ranges from a host buffer into a vector of device buffers
- *
- * @param host_buffer Host buffer span
- * @param byte_ranges Byte ranges to fetch
- * @param stream CUDA stream
- * @param mr Device memory resource to create device buffers with
- *
- * @return Vector of device buffers
- */
-std::vector<rmm::device_buffer> fetch_byte_ranges(
-  cudf::host_span<uint8_t const> host_buffer,
-  cudf::host_span<cudf::io::text::byte_range_info const> byte_ranges,
-  rmm::cuda_stream_view stream,
-  rmm::device_async_resource_ref mr)
-{
-  std::vector<rmm::device_buffer> buffers{};
-  buffers.reserve(byte_ranges.size());
-
-  std::transform(
-    byte_ranges.begin(),
-    byte_ranges.end(),
-    std::back_inserter(buffers),
-    [&](auto const& byte_range) {
-      auto const chunk_offset = host_buffer.data() + byte_range.offset();
-      auto const chunk_size   = byte_range.size();
-      auto buffer             = rmm::device_buffer(chunk_size, stream, mr);
-      CUDF_CUDA_TRY(cudaMemcpyAsync(
-        buffer.data(), chunk_offset, chunk_size, cudaMemcpyHostToDevice, stream.value()));
-      return buffer;
-    });
-
-  stream.synchronize_no_throw();
-  return buffers;
-}
-
-/**
- * @brief Creates a strings column with a constant stringified value between 0 and 9999
- *
- * @param value String value between 0 and 9999
- * @return Strings column wrapper
- */
-cudf::test::strings_column_wrapper constant_strings(cudf::size_type value)
-{
-  CUDF_EXPECTS(value >= 0 && value <= 9999, "String value must be between 0000 and 9999");
-
-  std::array<char, 5> buf;
-  auto elements =
-    thrust::make_transform_iterator(thrust::make_constant_iterator(value), [&buf](auto i) {
-      sprintf(buf.data(), "%04d", i);
-      return std::string(buf.data());
-    });
-  return cudf::test::strings_column_wrapper(elements, elements + num_ordered_rows);
-}
-
-/**
- * @brief Fail for types other than duration or timestamp
- */
-template <typename T, CUDF_ENABLE_IF(not cudf::is_chrono<T>())>
-cudf::test::fixed_width_column_wrapper<T> descending_low_cardinality()
-{
-  static_assert(
-    cudf::is_chrono<T>(),
-    "Use testdata::descending<T>() to generate descending values for non-temporal types");
-}
-
-/**
- * @brief Creates a duration column wrapper with low cardinality descending values
- *
- * @tparam T Duration type
- * @return Column wrapper
- */
-template <typename T, CUDF_ENABLE_IF(cudf::is_duration<T>())>
-cudf::test::fixed_width_column_wrapper<T> descending_low_cardinality()
-{
-  auto elements = cudf::detail::make_counting_transform_iterator(
-    0, [](auto i) { return T((num_ordered_rows - i) / 100); });
-  return cudf::test::fixed_width_column_wrapper<T>(elements, elements + num_ordered_rows);
-}
-
-/**
- * @brief Creates a timestamp column wrapper with low cardinality descending values
- *
- * @tparam T Timestamp type
- * @return Column wrapper
- */
-template <typename T, CUDF_ENABLE_IF(cudf::is_timestamp<T>())>
-cudf::test::fixed_width_column_wrapper<T> descending_low_cardinality()
-{
-  auto elements = cudf::detail::make_counting_transform_iterator(
-    0, [](auto i) { return T(typename T::duration((num_ordered_rows - i) / 100)); });
-  return cudf::test::fixed_width_column_wrapper<T>(elements, elements + num_ordered_rows);
-}
-
-/**
- * @brief Creates a table and writes it to Parquet host buffer with column level statistics
- *
- * This function creates a table with three columns:
- * - col0: ascending T values
- * - col1: descending T values (reduced cardinality for timestamps and durations)
- * - col2: constant cudf::string_view values
- *
- * The function creates a table by concatenating the same set of columns NumTableConcats times.
- * It then writes this table to a Parquet host buffer with column level statistics.
- *
- * @tparam T Data type for columns 0 and 1
- * @tparam NumTableConcats Number of times to concatenate the base table (must be >= 1)
- * @tparam IsConstantStrings Whether to use constant strings for column 2
- * @return Tuple of table and Parquet host buffer
- */
-template <typename T, size_t NumTableConcats, bool IsConstantStrings = true>
-auto create_parquet_with_stats(
-  cudf::size_type col2_value             = 100,
-  cudf::io::compression_type compression = cudf::io::compression_type::AUTO,
-  rmm::cuda_stream_view stream           = cudf::get_default_stream())
-{
-  static_assert(NumTableConcats >= 1, "Concatenated table must contain at least one table");
-
-  auto col0 = testdata::ascending<T>();
-  auto col1 = []() {
-    if constexpr (cudf::is_chrono<T>()) {
-      return descending_low_cardinality<T>();
-    } else {
-      return testdata::descending<T>();
-    }
-  }();
-
-  auto col2 = [&]() {
-    if constexpr (IsConstantStrings) {
-      return constant_strings(col2_value);  // constant stringified value
-    } else {
-      return testdata::ascending<cudf::string_view>();  // ascending strings
-    }
-  }();
-
-  auto expected = table_view{{col0, col1, col2}};
-  auto table    = cudf::concatenate(std::vector<table_view>(NumTableConcats, expected));
-  expected      = table->view();
-
-  cudf::io::table_input_metadata expected_metadata(expected);
-  expected_metadata.column_metadata[0].set_name("col0");
-  expected_metadata.column_metadata[1].set_name("col1");
-  expected_metadata.column_metadata[2].set_name("col2");
-
-  std::vector<char> buffer;
-  cudf::io::parquet_writer_options out_opts =
-    cudf::io::parquet_writer_options::builder(cudf::io::sink_info{&buffer}, expected)
-      .metadata(std::move(expected_metadata))
-      .row_group_size_rows(page_size_for_ordered_tests)
-      .max_page_size_rows(page_size_for_ordered_tests / 5)
-      .compression(compression)
-      .dictionary_policy(cudf::io::dictionary_policy::ALWAYS)
-      .stats_level(cudf::io::statistics_freq::STATISTICS_COLUMN);
-
-  if constexpr (NumTableConcats > 1) {
-    out_opts.set_row_group_size_rows(num_ordered_rows);
-    out_opts.set_max_page_size_rows(page_size_for_ordered_tests);
-  }
-
-  cudf::io::write_parquet(out_opts);
-
-  return std::pair{std::move(table), std::move(buffer)};
-}
 
 /**
  * @brief Filter input row groups using column chunk dictionaries via the experimental parquet
@@ -316,76 +95,10 @@ auto filter_row_groups_with_dictionaries(cudf::host_span<uint8_t const> file_buf
 
 }  // namespace
 
-TEST_F(ParquetExperimentalReaderTest, TestMetadata)
-{
-  srand(0xf00d);
-  using T = uint32_t;
+// Base test fixture for tests
+struct HybridScanPredicatePushdownTest : public cudf::test::BaseFixture {};
 
-  // Create a table with several row groups each with a single page.
-  auto constexpr num_concat         = 1;
-  auto constexpr rows_per_row_group = page_size_for_ordered_tests;
-  auto file_buffer                  = std::get<1>(create_parquet_with_stats<T, num_concat>());
-
-  // Filtering AST - table[0] < 100
-  auto literal_value     = cudf::numeric_scalar<T>(100);
-  auto literal           = cudf::ast::literal(literal_value);
-  auto col_ref_0         = cudf::ast::column_name_reference("col0");
-  auto filter_expression = cudf::ast::operation(cudf::ast::ast_operator::LESS, col_ref_0, literal);
-
-  // Create reader options with empty source info
-  cudf::io::parquet_reader_options options =
-    cudf::io::parquet_reader_options::builder().filter(filter_expression);
-
-  // Input file buffer span
-  auto const file_buffer_span = cudf::host_span<uint8_t const>(
-    reinterpret_cast<uint8_t const*>(file_buffer.data()), file_buffer.size());
-
-  // Fetch footer and page index bytes from the buffer.
-  auto const footer_buffer = fetch_footer_bytes(file_buffer_span);
-
-  // Create hybrid scan reader with footer bytes
-  auto const reader =
-    std::make_unique<cudf::io::parquet::experimental::hybrid_scan_reader>(footer_buffer, options);
-
-  // Get Parquet file metadata from the reader
-  auto parquet_metadata = reader->parquet_metadata();
-
-  // Check that the offset and column indices are not present
-  ASSERT_FALSE(parquet_metadata.row_groups[0].columns[0].offset_index.has_value());
-  ASSERT_FALSE(parquet_metadata.row_groups[0].columns[0].column_index.has_value());
-
-  // Get page index byte range from the reader
-  auto const page_index_byte_range = reader->page_index_byte_range();
-
-  // Fetch page index bytes from the input buffer
-  auto const page_index_buffer = fetch_page_index_bytes(file_buffer_span, page_index_byte_range);
-
-  // Setup page index
-  reader->setup_page_index(page_index_buffer);
-
-  // Get Parquet file metadata from the reader again
-  parquet_metadata = reader->parquet_metadata();
-
-  // Check that the offset and column indices are now present
-  ASSERT_TRUE(parquet_metadata.row_groups[0].columns[0].offset_index.has_value());
-  ASSERT_TRUE(parquet_metadata.row_groups[0].columns[0].column_index.has_value());
-
-  // Get all row groups from the reader
-  auto input_row_group_indices = reader->all_row_groups(options);
-  // Expect 4 = 20000 rows / 5000 rows per row group
-  EXPECT_EQ(input_row_group_indices.size(), 4);
-
-  // Explicitly set the row groups to read
-  options.set_row_groups({{0, 1}});
-
-  // Get all row groups from the reader again
-  input_row_group_indices = reader->all_row_groups(options);
-  // Expect only 2 row groups now
-  EXPECT_EQ(input_row_group_indices.size(), 2);
-  EXPECT_EQ(reader->total_rows_in_row_groups(input_row_group_indices), 2 * rows_per_row_group);
-}
-
-TEST_F(ParquetExperimentalReaderTest, TestFilterRowGroupsWithStats)
+TEST_F(HybridScanPredicatePushdownTest, FilterRowGroupsWithStats)
 {
   srand(0xc001);
   using T = uint32_t;
@@ -438,7 +151,7 @@ TEST_F(ParquetExperimentalReaderTest, TestFilterRowGroupsWithStats)
 }
 
 template <typename T>
-struct PageFilteringWithPageIndexStats : public ParquetExperimentalReaderTest {};
+struct PageFilteringWithPageIndexStats : public HybridScanPredicatePushdownTest {};
 
 // Unsigned numeric types except booleans for columns 0 and 1 for page index stats tests
 using SignedIntegralTypesNotBool =
@@ -449,7 +162,7 @@ using PageFilteringTestTypes =
 
 TYPED_TEST_SUITE(PageFilteringWithPageIndexStats, PageFilteringTestTypes);
 
-TYPED_TEST(PageFilteringWithPageIndexStats, TestFilterPagesWithPageIndexStats)
+TYPED_TEST(PageFilteringWithPageIndexStats, FilterPagesWithPageIndexStats)
 {
   using T = TypeParam;
 
@@ -601,7 +314,7 @@ TYPED_TEST(PageFilteringWithPageIndexStats, TestFilterPagesWithPageIndexStats)
   }
 }
 
-TEST_F(ParquetExperimentalReaderTest, TestFilterRowGroupsWithDictBasic)
+TEST_F(HybridScanPredicatePushdownTest, FilterRowGroupsWithDictBasic)
 {
   srand(0xcafe);
   using T = uint32_t;
@@ -878,7 +591,7 @@ TEST_F(ParquetExperimentalReaderTest, TestFilterRowGroupsWithDictBasic)
 }
 
 template <typename T>
-struct RowGroupFilteringWithDictTest : public ParquetExperimentalReaderTest {};
+struct RowGroupFilteringWithDictTest : public HybridScanPredicatePushdownTest {};
 
 // Booleans are not supported for dictionary based filtering
 using DictionaryTestTypes =
