+++ cpp/tests/io/experimental/hybrid_scan_test.cpp
@@ -14,21 +14,17 @@
  * limitations under the License.
  */
 
-#include "parquet_common.hpp"
+#include "hybrid_scan_common.hpp"
 
 #include <cudf_test/base_fixture.hpp>
-#include <cudf_test/column_wrapper.hpp>
 #include <cudf_test/io_metadata_utilities.hpp>
-#include <cudf_test/iterator_utilities.hpp>
 #include <cudf_test/table_utilities.hpp>
 
 #include <cudf/column/column.hpp>
-#include <cudf/concatenate.hpp>
 #include <cudf/io/experimental/hybrid_scan.hpp>
 #include <cudf/io/parquet.hpp>
 #include <cudf/io/text/byte_range_info.hpp>
 #include <cudf/stream_compaction.hpp>
-#include <cudf/table/table.hpp>
 #include <cudf/table/table_view.hpp>
 #include <cudf/transform.hpp>
 #include <cudf/utilities/default_stream.hpp>
@@ -37,142 +33,9 @@
 
 #include <rmm/mr/device/aligned_resource_adaptor.hpp>
 
-// Base test fixture for tests
-struct ParquetExperimentalReaderTest : public cudf::test::BaseFixture {};
 auto constexpr bloom_filter_alignment = 32;
 
 namespace {
-
-/**
- * @brief Fetches a host span of Parquet footer bytes from the input buffer span
- *
- * @param buffer Input buffer span
- * @return A host span of the footer bytes
- */
-cudf::host_span<uint8_t const> fetch_footer_bytes(cudf::host_span<uint8_t const> buffer)
-{
-  using namespace cudf::io::parquet;
-
-  constexpr auto header_len = sizeof(file_header_s);
-  constexpr auto ender_len  = sizeof(file_ender_s);
-  size_t const len          = buffer.size();
-
-  auto const header_buffer = cudf::host_span<uint8_t const>(buffer.data(), header_len);
-  auto const header        = reinterpret_cast<file_header_s const*>(header_buffer.data());
-  auto const ender_buffer =
-    cudf::host_span<uint8_t const>(buffer.data() + len - ender_len, ender_len);
-  auto const ender = reinterpret_cast<file_ender_s const*>(ender_buffer.data());
-  CUDF_EXPECTS(len > header_len + ender_len, "Incorrect data source");
-  constexpr uint32_t parquet_magic = (('P' << 0) | ('A' << 8) | ('R' << 16) | ('1' << 24));
-  CUDF_EXPECTS(header->magic == parquet_magic && ender->magic == parquet_magic,
-               "Corrupted header or footer");
-  CUDF_EXPECTS(ender->footer_len != 0 && ender->footer_len <= (len - header_len - ender_len),
-               "Incorrect footer length");
-
-  return cudf::host_span<uint8_t const>(buffer.data() + len - ender->footer_len - ender_len,
-                                        ender->footer_len);
-}
-
-/**
- * @brief Fetches a host span of Parquet PageIndexbytes from the input buffer span
- *
- * @param buffer Input buffer span
- * @param page_index_bytes Byte range of `PageIndex` to fetch
- * @return A host span of the PageIndex bytes
- */
-cudf::host_span<uint8_t const> fetch_page_index_bytes(
-  cudf::host_span<uint8_t const> buffer, cudf::io::text::byte_range_info const page_index_bytes)
-{
-  return cudf::host_span<uint8_t const>(
-    reinterpret_cast<uint8_t const*>(buffer.data()) + page_index_bytes.offset(),
-    page_index_bytes.size());
-}
-
-/**
- * @brief Fetches a list of byte ranges from a host buffer into a vector of device buffers
- *
- * @param host_buffer Host buffer span
- * @param byte_ranges Byte ranges to fetch
- * @param stream CUDA stream
- * @param mr Device memory resource to create device buffers with
- *
- * @return Vector of device buffers
- */
-std::vector<rmm::device_buffer> fetch_byte_ranges(
-  cudf::host_span<uint8_t const> host_buffer,
-  cudf::host_span<cudf::io::text::byte_range_info const> byte_ranges,
-  rmm::cuda_stream_view stream,
-  rmm::device_async_resource_ref mr)
-{
-  std::vector<rmm::device_buffer> buffers{};
-  buffers.reserve(byte_ranges.size());
-
-  std::transform(
-    byte_ranges.begin(),
-    byte_ranges.end(),
-    std::back_inserter(buffers),
-    [&](auto const& byte_range) {
-      auto const chunk_offset = host_buffer.data() + byte_range.offset();
-      auto const chunk_size   = byte_range.size();
-      auto buffer             = rmm::device_buffer(chunk_size, stream, mr);
-      CUDF_CUDA_TRY(cudaMemcpyAsync(
-        buffer.data(), chunk_offset, chunk_size, cudaMemcpyHostToDevice, stream.value()));
-      return buffer;
-    });
-
-  stream.synchronize_no_throw();
-  return buffers;
-}
-
-/**
- * @brief Creates a table and writes it to Parquet host buffer with column level statistics
- *
- * This function creates a table with three columns:
- * - col_uint32: ascending uint32_t values
- * - col_int64: descending int64_t values
- * - col_str: ascending string values
- *
- * The function creates a table by concatenating the same set of columns NumTableConcats times.
- * It then writes this table to a Parquet host buffer with column level statistics.
- *
- * @tparam NumTableConcats Number of times to concatenate the base table (must be >= 1)
- * @return Tuple of table and Parquet host buffer
- */
-template <size_t NumTableConcats>
-auto create_parquet_with_stats()
-{
-  static_assert(NumTableConcats >= 1, "Concatenated table must contain at least one table");
-
-  auto col0 = testdata::ascending<uint32_t>();
-  auto col1 = testdata::descending<int64_t>();
-  auto col2 = testdata::ascending<cudf::string_view>();
-
-  auto expected = table_view{{col0, col1, col2}};
-  auto table    = cudf::concatenate(std::vector<table_view>(NumTableConcats, expected));
-  expected      = table->view();
-
-  cudf::io::table_input_metadata expected_metadata(expected);
-  expected_metadata.column_metadata[0].set_name("col_uint32");
-  expected_metadata.column_metadata[1].set_name("col_int64");
-  expected_metadata.column_metadata[2].set_name("col_str");
-
-  std::vector<char> buffer;
-  cudf::io::parquet_writer_options out_opts =
-    cudf::io::parquet_writer_options::builder(cudf::io::sink_info{&buffer}, expected)
-      .metadata(std::move(expected_metadata))
-      .row_group_size_rows(5000)
-      .max_page_size_rows(1000)
-      .stats_level(cudf::io::statistics_freq::STATISTICS_COLUMN);
-
-  if constexpr (NumTableConcats > 1) {
-    out_opts.set_row_group_size_rows(20000);
-    out_opts.set_max_page_size_rows(5000);
-  }
-
-  cudf::io::write_parquet(out_opts);
-
-  return std::pair{std::move(table), std::move(buffer)};
-}
 
 /**
  * @brief Read parquet file with the hybrid scan reader
@@ -325,21 +188,94 @@ auto hybrid_scan(std::vector<char>& buffer,
 
 }  // namespace
 
-TEST_F(ParquetExperimentalReaderTest, PruneRowGroupsOnly)
+// Base test fixture for tests
+struct HybridScanTest : public cudf::test::BaseFixture {};
+
+TEST_F(HybridScanTest, TestMetadata)
 {
-  srand(31337);
+  srand(0xf00d);
+  using T = uint32_t;
+
+  // Create a table with several row groups each with a single page.
+  auto constexpr num_concat         = 1;
+  auto constexpr rows_per_row_group = page_size_for_ordered_tests;
+  auto file_buffer                  = std::get<1>(create_parquet_with_stats<T, num_concat>());
+
+  // Filtering AST - table[0] < 100
+  auto literal_value     = cudf::numeric_scalar<T>(100);
+  auto literal           = cudf::ast::literal(literal_value);
+  auto col_ref_0         = cudf::ast::column_name_reference("col0");
+  auto filter_expression = cudf::ast::operation(cudf::ast::ast_operator::LESS, col_ref_0, literal);
+
+  // Create reader options with empty source info
+  cudf::io::parquet_reader_options options =
+    cudf::io::parquet_reader_options::builder().filter(filter_expression);
+
+  // Input file buffer span
+  auto const file_buffer_span = cudf::host_span<uint8_t const>(
+    reinterpret_cast<uint8_t const*>(file_buffer.data()), file_buffer.size());
+
+  // Fetch footer and page index bytes from the buffer.
+  auto const footer_buffer = fetch_footer_bytes(file_buffer_span);
+
+  // Create hybrid scan reader with footer bytes
+  auto const reader =
+    std::make_unique<cudf::io::parquet::experimental::hybrid_scan_reader>(footer_buffer, options);
+
+  // Get Parquet file metadata from the reader
+  auto parquet_metadata = reader->parquet_metadata();
+
+  // Check that the offset and column indices are not present
+  ASSERT_FALSE(parquet_metadata.row_groups[0].columns[0].offset_index.has_value());
+  ASSERT_FALSE(parquet_metadata.row_groups[0].columns[0].column_index.has_value());
+
+  // Get page index byte range from the reader
+  auto const page_index_byte_range = reader->page_index_byte_range();
+
+  // Fetch page index bytes from the input buffer
+  auto const page_index_buffer = fetch_page_index_bytes(file_buffer_span, page_index_byte_range);
+
+  // Setup page index
+  reader->setup_page_index(page_index_buffer);
+
+  // Get Parquet file metadata from the reader again
+  parquet_metadata = reader->parquet_metadata();
+
+  // Check that the offset and column indices are now present
+  ASSERT_TRUE(parquet_metadata.row_groups[0].columns[0].offset_index.has_value());
+  ASSERT_TRUE(parquet_metadata.row_groups[0].columns[0].column_index.has_value());
+
+  // Get all row groups from the reader
+  auto input_row_group_indices = reader->all_row_groups(options);
+  // Expect 4 = 20000 rows / 5000 rows per row group
+  EXPECT_EQ(input_row_group_indices.size(), 4);
+
+  // Explicitly set the row groups to read
+  options.set_row_groups({{0, 1}});
+
+  // Get all row groups from the reader again
+  input_row_group_indices = reader->all_row_groups(options);
+  // Expect only 2 row groups now
+  EXPECT_EQ(input_row_group_indices.size(), 2);
+  EXPECT_EQ(reader->total_rows_in_row_groups(input_row_group_indices), 2 * rows_per_row_group);
+}
+
+TEST_F(HybridScanTest, PruneRowGroupsOnlyAndScanAllColumns)
+{
+  srand(0xc0ffee);
+  using T = uint32_t;
 
   // A table not concated with itself with result in a parquet file with several row groups each
   // with a single page. Since there is only one page per row group, the page and row group stats
   // are identical and we can only prune row groups.
   auto constexpr num_concat            = 1;
-  auto [written_table, parquet_buffer] = create_parquet_with_stats<num_concat>();
+  auto [written_table, parquet_buffer] = create_parquet_with_stats<T, num_concat>();
 
   // Filtering AST - table[0] < 100
   auto constexpr num_filter_columns = 1;
   auto literal_value                = cudf::numeric_scalar<uint32_t>(100);
   auto literal                      = cudf::ast::literal(literal_value);
-  auto col_ref_0                    = cudf::ast::column_name_reference("col_uint32");
+  auto col_ref_0                    = cudf::ast::column_name_reference("col0");
   auto filter_expression = cudf::ast::operation(cudf::ast::ast_operator::LESS, col_ref_0, literal);
 
   auto stream     = cudf::get_default_stream();
@@ -367,21 +303,22 @@ TEST_F(ParquetExperimentalReaderTest, PruneRowGroupsOnly)
   }
 }
 
-TEST_F(ParquetExperimentalReaderTest, TestPayloadColumns)
+TEST_F(HybridScanTest, PruneRowGroupsOnlyAndScanSelectColumns)
 {
-  srand(31337);
+  srand(0xcafe);
+  using T = cudf::timestamp_ms;
 
   // A table not concated with itself with result in a parquet file with several row groups each
   // with a single page. Since there is only one page per row group, the page and row group stats
   // are identical and we can only prune row groups.
   auto constexpr num_concat            = 1;
-  auto [written_table, parquet_buffer] = create_parquet_with_stats<num_concat>();
+  auto [written_table, parquet_buffer] = create_parquet_with_stats<T, num_concat>();
 
   // Filtering AST - table[0] < 100
   auto constexpr num_filter_columns = 1;
-  auto literal_value                = cudf::numeric_scalar<uint32_t>(100);
+  auto literal_value                = cudf::timestamp_scalar<T>(T{typename T::duration{100}});
   auto literal                      = cudf::ast::literal(literal_value);
-  auto col_ref_0                    = cudf::ast::column_name_reference("col_uint32");
+  auto col_ref_0                    = cudf::ast::column_name_reference("col0");
   auto filter_expression = cudf::ast::operation(cudf::ast::ast_operator::LESS, col_ref_0, literal);
 
   auto stream     = cudf::get_default_stream();
@@ -390,7 +327,7 @@ TEST_F(ParquetExperimentalReaderTest, TestPayloadColumns)
     cudf::get_current_device_resource(), bloom_filter_alignment);
 
   {
-    auto const payload_column_names = std::vector<std::string>{"col_uint32", "col_str"};
+    auto const payload_column_names = std::vector<std::string>{"col0", "col2"};
     // Read parquet using the hybrid scan reader
     auto [read_filter_table, read_payload_table, read_filter_meta, read_payload_meta, row_mask] =
       hybrid_scan(parquet_buffer,
@@ -413,7 +350,7 @@ TEST_F(ParquetExperimentalReaderTest, TestPayloadColumns)
   }
 
   {
-    auto const payload_column_names = std::vector<std::string>{"col_str", "col_int64"};
+    auto const payload_column_names = std::vector<std::string>{"col2", "col1"};
     // Read parquet using the hybrid scan reader
     auto [read_filter_table, read_payload_table, read_filter_meta, read_payload_meta, row_mask] =
       hybrid_scan(parquet_buffer,
@@ -436,21 +373,22 @@ TEST_F(ParquetExperimentalReaderTest, TestPayloadColumns)
   }
 }
 
-TEST_F(ParquetExperimentalReaderTest, PrunePagesOnly)
+TEST_F(HybridScanTest, PruneDataPagesOnlyAndScanAllColumns)
 {
-  srand(31337);
+  srand(0xf00d);
+  using T = cudf::duration_ms;
 
   // A table concatenated multiple times by itself with result in a parquet file with a row group
-  // per concatenation with multiple pages per row group. Since all row groups will be identical, we
-  // can only prune pages based on `PageIndex` stats
+  // per concatenation with multiple pages per row group. Since all row groups will be identical,
+  // we can only prune pages based on `PageIndex` stats
   auto constexpr num_concat    = 2;
-  auto [written_table, buffer] = create_parquet_with_stats<num_concat>();
+  auto [written_table, buffer] = create_parquet_with_stats<T, num_concat>();
 
   // Filtering AST - table[0] < 100
   auto constexpr num_filter_columns = 1;
-  auto literal_value                = cudf::numeric_scalar<uint32_t>(100);
+  auto literal_value                = cudf::duration_scalar<T>(T{100});
   auto literal                      = cudf::ast::literal(literal_value);
-  auto col_ref_0                    = cudf::ast::column_name_reference("col_uint32");
+  auto col_ref_0                    = cudf::ast::column_name_reference("col0");
   auto filter_expression = cudf::ast::operation(cudf::ast::ast_operator::LESS, col_ref_0, literal);
 
   auto stream     = cudf::get_default_stream();
